backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6

faiss-cpu==1.8.0.post1

pypdf==4.2.0
python-docx==1.1.2

requests==2.32.3

pydantic==2.9.2
pydantic-settings==2.5.2

python-multipart==0.0.9
